Move Compress and Uncompress to compression.go
